Use font-awesome everywhere (#232)
use font-awesome everywhere

Now, using 2 font families of font-awesome and font-glyphicon. And
bootstrap4 stopped to include font-glyphicon, so this uses
font-awesome everywhere.

To do it,

- add dependency to xstatic-font-awesome, instead of fetching from internet.
- replace icon in font-glyphicon to similar icon in font-awesome
  [glyphicon-heart is replaced with fa-unlock, because font-awesome
   has real unlock icon.]

With this changes, user doesn't need to download 2 fonts, and prepare
for bootstrap4.

Also:

Fix indent and untabify for templates/index.html

Use 4 indent, and stop mixing spc and tab.