add +thumbnail view, fixes #177
same usage as +download or +inline, but dynamically computes a thumbnail in memory.

this function is not yet exposed on the UI, because it is intended to be used
by some future "image gallery" feature.

for thumbnail computation, Pillow (PIL fork) is used, which can be installed
as an extra, but optional requirement.