Use setup.cfg instead of setup.py
Besides fully migrating to setup.cfg, the versioning system is also
changed from setuptools_scm to a simple attribute in `bepasty/__init__.py`.

This would require manual editing of the `bepasty/__init__.py` in order to
control the versioning from now on.

Removing setuptools_scm as a dependency will allow the package to be
built without a .git folder present, which is favorable for packaging
a docker image, otherwise `pip install -e .` would fail.